+++ SRC/GekkoCore/Interpreter/Condition.cpp
@@ -8,7 +8,7 @@ namespace Gekko
 	// CR[crbd] = CR[crba] & CR[crbb]
 	void Interpreter::crand()
 	{
-		uint32_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
+		size_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
 		
 		uint32_t a = (core->regs.cr >> (31 - crba)) & 1;
 		uint32_t b = (core->regs.cr >> (31 - crbb)) & 1;
@@ -21,7 +21,7 @@ namespace Gekko
 	// CR[crbd] = CR[crba] & ~CR[crbb]
 	void Interpreter::crandc()
 	{
-		uint32_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
+		size_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
 
 		uint32_t a = (core->regs.cr >> (31 - crba)) & 1;
 		uint32_t b = (core->regs.cr >> (31 - crbb)) & 1;
@@ -34,7 +34,7 @@ namespace Gekko
 	// CR[crbd] = CR[crba] EQV CR[crbb]
 	void Interpreter::creqv()
 	{
-		uint32_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
+		size_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
 
 		uint32_t a = (core->regs.cr >> (31 - crba)) & 1;
 		uint32_t b = (core->regs.cr >> (31 - crbb)) & 1;
@@ -47,7 +47,7 @@ namespace Gekko
 	// CR[crbd] = !(CR[crba] & CR[crbb])
 	void Interpreter::crnand()
 	{
-		uint32_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
+		size_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
 
 		uint32_t a = (core->regs.cr >> (31 - crba)) & 1;
 		uint32_t b = (core->regs.cr >> (31 - crbb)) & 1;
@@ -60,7 +60,7 @@ namespace Gekko
 	// CR[crbd] = !(CR[crba] | CR[crbb])
 	void Interpreter::crnor()
 	{
-		uint32_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
+		size_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
 
 		uint32_t a = (core->regs.cr >> (31 - crba)) & 1;
 		uint32_t b = (core->regs.cr >> (31 - crbb)) & 1;
@@ -73,7 +73,7 @@ namespace Gekko
 	// CR[crbd] = CR[crba] | CR[crbb]
 	void Interpreter::cror()
 	{
-		uint32_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
+		size_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
 
 		uint32_t a = (core->regs.cr >> (31 - crba)) & 1;
 		uint32_t b = (core->regs.cr >> (31 - crbb)) & 1;
@@ -86,7 +86,7 @@ namespace Gekko
 	// CR[crbd] = CR[crba] | ~CR[crbb]
 	void Interpreter::crorc()
 	{
-		uint32_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
+		size_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
 
 		uint32_t a = (core->regs.cr >> (31 - crba)) & 1;
 		uint32_t b = (core->regs.cr >> (31 - crbb)) & 1;
@@ -99,7 +99,7 @@ namespace Gekko
 	// CR[crbd] = CR[crba] ^ CR[crbb]
 	void Interpreter::crxor()
 	{
-		uint32_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
+		size_t crbd = info.paramBits[0], crba = info.paramBits[1], crbb = info.paramBits[2];
 
 		uint32_t a = (core->regs.cr >> (31 - crba)) & 1;
 		uint32_t b = (core->regs.cr >> (31 - crbb)) & 1;
@@ -112,7 +112,7 @@ namespace Gekko
 	// CR[4*crfd .. 4*crfd + 3] = CR[4*crfs .. 4*crfs + 3]
 	void Interpreter::mcrf()
 	{
-		int32_t crfd = 4 * (7 - info.paramBits[0]), crfs = 4 * (7 - info.paramBits[1]);
+		size_t crfd = 4 * (7 - info.paramBits[0]), crfs = 4 * (7 - info.paramBits[1]);
 		core->regs.cr = (core->regs.cr & (~(0xf << crfd))) | (((core->regs.cr >> crfs) & 0xf) << crfd);
 		core->regs.pc += 4;
 	}
